requirements.txt
django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.10.0
Pillow==8.4.0
